% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLocationTimeDate.R
\name{add.date}
\alias{add.date}
\title{Add date information to data.table}
\usage{
add.date(x)
}
\arguments{
\item{x}{The dataframe wiuth detection data, which should have filepath as a
column and the file names should have adopted convention defined in nameFiles}
}
\description{
The following functions use the file nameing convention established earlier
to add relevant information (date, time, and location) to the detection dataframe.
It can take some time to run depending on size of dataset.
}
