% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLocationTimeDate.R
\name{add.locationID}
\alias{add.locationID}
\title{Add location information to data.table}
\usage{
add.locationID(x)
}
\arguments{
\item{x}{The dataframe wiuth detection data, which should have filepath as a
column and the file names should have adopted convention defined in nameFiles}
}
\description{
Add location information to data.table
}
