% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLocationTimeDate.R
\name{add.time}
\alias{add.time}
\title{Add time information to data.table}
\usage{
add.time(x)
}
\arguments{
\item{x}{The dataframe wiuth detection data, which should have filepath as a
column and the file names should have adopted convention defined in nameFiles}
}
\description{
Add time information to data.table
}
