% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWaves.R
\name{makeWaves}
\alias{makeWaves}
\title{Download audio files corresponding to detection timestamps}
\usage{
makeWaves(x, dir, time_buffer = 0, sampling_rate = 48000)
}
\arguments{
\item{x}{A dataframe of detections with filepath, start, and end columns.}

\item{dir}{The directory to download audio files to.}

\item{time_buffer}{Defines how many seconds around each detection window to
include in downloaded files.}

\item{sampling_rate}{Defines the desired sampling rate for files. Defaults to 48000.}
}
\description{
This function downloads audio files to a directory that correspond to detection
timestamps in a detection dataframe. This is useful for downloading subsets for
verification or sharing.
}
