% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameFiles.R
\name{nameFiles}
\alias{nameFiles}
\title{This will name all files in a directory (usually SD card) according to the
arguments of locID and eqID so that "locationID_deviceID_YYYYMMDD_HHMMSS.WAV"
becomes the naming convention. This will help a great deal with organization
and later analysis, since each filename is encoded with a lot of crucial
information. The original files are not deleted so that you can verify successful
copying before wiping sd card clean.}
\usage{
nameFiles(directory, new_directory, locID, eqID)
}
\arguments{
\item{directory}{Character string indicating the directory where the files are}

\item{new_directory}{Character string indicating the new directory where renamed
files should be copied. This should be created ahead of time manually.}

\item{locID}{Character string indicating the location or site ID where recordings
were collected from.}

\item{eqID}{Character string indicating the equipment ID}
}
\value{
Nothing
}
\description{
This will name all files in a directory (usually SD card) according to the
arguments of locID and eqID so that "locationID_deviceID_YYYYMMDD_HHMMSS.WAV"
becomes the naming convention. This will help a great deal with organization
and later analysis, since each filename is encoded with a lot of crucial
information. The original files are not deleted so that you can verify successful
copying before wiping sd card clean.
}
