% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_specs.R
\name{plot_specs}
\alias{plot_specs}
\title{This function will plot spectrorams in a grid for quick review.}
\usage{
plot_specs(x, n_specs = 6)
}
\description{
@param x a dataframe of detections with columns for common_name and confidence.
@param n_specs Number of spectrograms you'd like to plot. Defaults to 5.
@return No return
@importFrom monitoR viewSpec
@export
}
