% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullWave.R
\name{pullWave}
\alias{pullWave}
\title{Extract and play a wave segment from an audio file based on detection data}
\usage{
pullWave(
  data,
  det_num,
  time.buffer = 0,
  start_buffer = 0,
  end_buffer = 0,
  listen = TRUE
)
}
\arguments{
\item{data}{A data frame containing detection information with columns 'filepath', 'start', and 'end'.}

\item{det_num}{An integer representing the index of the detection in the data frame.}

\item{time.buffer}{A numeric value specifying the time buffer (in seconds) to add before and after the detected segment (default: 0).}

\item{start_buffer}{A numeric value specifying the additional time buffer (in seconds) to add before the detected segment (default: 0).}

\item{end_buffer}{A numeric value specifying the additional time buffer (in seconds) to add after the detected segment (default: 0).}

\item{listen}{A logical value indicating whether to play the extracted audio segment (default: TRUE).}
}
\value{
A 'Wave' object representing the extracted audio segment, including the specified time buffers.
}
\description{
This function reads an audio file and extracts a segment of the audio based on the provided detection data.
The extracted segment can be played, and a spectrogram is displayed.
}
\examples{
\dontrun{
  # Assuming 'detection_data' is a data frame with the required columns
  extracted_wave <- pullWave(data = detection_data, det_num = 1)
}
}
