% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_count.R
\name{species_count}
\alias{species_count}
\title{This function creates summary statistics of detection dataframes. It will show
species count, the maximum and minimum confidence detection per species, the mean
and median confidence scores per species, the standard deviation of confidence
scores, as well as a 'high confidence mean' value which is simply the average
confidence score for the n highest confidence detection, where n is defined by
the num.obs argument.}
\usage{
species_count(data, num.obs = 20)
}
\description{
@param data a dataframe of detections with columns for common_name and confidence.
@param num.obs The number of observations to consider when calculating 'high
still finnicky and is only useful if labeling data for model creation.
@return spec.table, which is a table with species, counts, and summary
statistics for confidence values
@export
}
