% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_results.R
\name{validate_results}
\alias{validate_results}
\title{Verify detection results}
\usage{
validate_results(
  data,
  species = "all",
  time_buffer = 0,
  conf = 0,
  temp.length = "none",
  template.mode = FALSE
)
}
\arguments{
\item{species}{A character string defining the species (as listed in BirdNET).}

\item{time_buffer}{Numeric, defines how much time around each detection window
you would like to visualize spectrograms.}

\item{conf}{Numeric, defining the lower threshold for detection confidence.}

\item{temp.length}{Numeric, defining the number of seconds each template should be
if using template mode.}

\item{template.mode}{Boolean, if set to TRUE, template mode is on. This is
still finnicky and is only useful if labeling data for model creation.}

\item{filepath}{A character string defining the filepath to the orginial audio file.}
}
\value{
Originial data.table with added columns 'verification' and 'notes'
with annotations from executing the function.
}
\description{
This function expects certain columns to exist and was made to work with
data.frames generated from BirdNET detections. The columns required are:
'filepath': a character string defining the filepath to the original wav file
'start': numeric, defining the start time of the detection window
'end': numeric, defining the end time of the detection window
'confidence': numeric, defining the confidence level
Optionally, you can have 'common_name' defining species
}
