% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_scan.R, R/verify_presence.R
\name{verify_presence}
\alias{verify_presence}
\alias{fast_scan}
\title{Fast-scan through detection results
This function expects certain columns to exist and was made to work with
data.frames generated from BirdNET detections. The columns required are:
'filepath': a character string defining the filepath to the original wav file
'start': numeric, defining the start time of the detection window
'end': numeric, defining the end time of the detection window
'confidence': numeric, defining the confidence level
'common_name': defining species}
\usage{
fast_scan(x, n_specs = 12)

verify_presence(data, species = "all", conf = 0)
}
\arguments{
\item{data}{a dataframe of detection results.}

\item{species}{A character string defining the species (as listed in BirdNET).}

\item{conf}{Numeric, defining the lower threshold for detection confidence.}
}
\value{
a new dataframe with all species in one column and a 1 or 0 if presence
was either confirmed or denied

a new dataframe with all species in one column and a 1 or 0 if presence
was either confirmed or denied
}
\description{
The function allows you to go scan quickly through detection results, plotting
multiple spectrograms at once with their common names and confidence levels.
The user input can either b 'c' to continue to next page, or a number corresponding
to the index value of the spectrogram you would like to hear.

The function allows you to go through your data and verify the presence of
each species. Once a 'y' is marked for one species, it moves onto the next
species. This can be very helpful if wanting a simple species richness
estimate with a certain dataset.
}
