% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_results.R
\name{verify_results}
\alias{verify_results}
\title{Verify detection results
This function expects certain columns to exist and was made to work with
data.frames generated from BirdNET detections. The columns required are:
'filepath': a character string defining the filepath to the original wav file
'start': numeric, defining the start time of the detection window
'end': numeric, defining the end time of the detection window
'confidence': numeric, defining the confidence level
Optionally, you can have 'common_name' defining species}
\usage{
verify_results(
  data,
  species = "all",
  time_buffer = 0,
  conf = 0,
  temp.length = "none",
  template.mode = FALSE
)
}
\description{
@param filepath A character string defining the filepath to the orginial audio file.
@param species A character string defining the species (as listed in BirdNET)
@param time_buffer Numeric, defines how much time around each detection window
you wouldl like to visualize spectrograms
@param conf Numeric, defining the lower threshold for detection confidence
@param temp.length Numeric, defining the number of seconds each template should be
if using template mode.
@param template.mode Boolean, if set to TRUE, template mode is on. This is
still finnicky and is only useful if labeling data for model creation.
@return original data.table with added columns 'verificaiton' and  'notes'
with annotations from executing the function
@import data.table
@importFrom tuneR readWave, writeWave
@importFrom seewave spectro
@importFrom monitoR viewSpec
}
\details{
@export
}
